#70 - Moved UUIDSerializer to json package
